advisor_ntwrk_project/requirements.txt
django
djangorestframework
pillow
djangorestframework-simplejwt
gunicorn